+++ src/app/(auth)/Sign-up.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 import Screen from '@/src/components/Screen'
 import SignupScreens from './SignUp/SignupScreens'
-import { UserData } from '@/src/Schemas/UserSchema'
+import { UserData } from '@/src/schemas/UserSchema'
 
 export default function SignUp() {
   const [step, setStep] = useState(5)
