+++ src/app/(auth)/Validate-otp.tsx
@@ -5,9 +5,9 @@ import { router } from 'expo-router'
 import { zodResolver } from '@hookform/resolvers/zod'
 import Icons from '@/src/components/Icons'
 import Screen from '@/src/components/Screen'
-import { OTPSchema } from '@/src/Schemas/UserSchema'
-import CustomInputs from '@/src/components/CustomInputs'
-import CustomButton from '@/src/components/CustomButton'
+import { OTPSchema } from '@/src/schemas/UserSchema'
+import Input from '@/src/components/Input'
+import Button from '@/src/components/Button'
 
 export default function RecoveryFunction() {
   const { HappyIcon } = Icons
@@ -25,12 +25,12 @@ export default function RecoveryFunction() {
             Hemos enviado el código a tu correo
           </Text>
           <Text className="text-lightc font-pbold text-[16px]">por favor verificalo</Text>
-          <CustomInputs variant="number" control={control} name="otp">
+          <Input variant="number" control={control} name="otp">
             Código
-          </CustomInputs>
+          </Input>
         </View>
         <View className="flex flex-row justify-around mt-20">
-          <CustomButton
+          <Button
             width={130}
             height={47}
             variant="secondary"
@@ -40,8 +40,8 @@ export default function RecoveryFunction() {
             }}
           >
             Cancelar
-          </CustomButton>
-          <CustomButton
+          </Button>
+          <Button
             width={130}
             height={47}
             variant="primary"
@@ -53,7 +53,7 @@ export default function RecoveryFunction() {
             })}
           >
             Siguiente
-          </CustomButton>
+          </Button>
         </View>
       </Screen>
     </ScrollView>
