--- src/components/navigation/TabBarIcon.tsx
@@ -1,7 +1,0 @@
-import React, { type ComponentProps } from 'react'
-import Ionicons from '@expo/vector-icons/Ionicons'
-import { type IconProps } from '@expo/vector-icons/build/createIconSet'
-
-export function TabBarIcon({ style, ...rest }: IconProps<ComponentProps<typeof Ionicons>['name']>) {
-  return <Ionicons size={28} style={[{ marginBottom: -3 }, style]} {...rest} />
-}
